Created copy object function for copying list content
It was required to have a function pointer parameter for passing a deep copy function to the copyContentToList() function. This is needed for being able to copy Object members from list element to list element by using deep copying. The default function does not copy anything and is required only for avoiding compile errors. The user is responsible to pass a custom deep copy function to ensure objects with references to dynamically allocated heap memory are correctly copied