Implemented memory pool for List elements
This was required in order to obtain better performance (cache locality) by
pre-allocating batches of ListElement objects, which would then be fed to
lists on demand. The elements would be returned to the pool (instead of
being deleted) once no longer required.

The pool can be shared among multiple lists which further improves
efficiency.

Additional changes:
- improved the linked list and collections tests by creating a cleanup
  mechanism that would be run after each test method
- improved the internal structure of the HashTable by allocating list buckets that are contiguous in memory