Implemented preventing double release to pool
It was required not to allow releasing the same ListElement more than
once to the ListElementsPool in order to prevent a future allocation to more than one
consumer (which would cause serious memory errors).

Some refactoring of the ListElementsPool and an update of its test were
also needed. The test would crash due to double free error when failing.
This fix was meant to prevent this from happening.